requirements.txt
gradio
python-dotenv
pydantic
google-generativeai
tavily-python
numpy
faiss-cpu
langchain
sentence-transformers